MDEV-174: Moved transformValidators to utils
