+++ packages/report-server/src/reportBuilder/transform/functions/updateColumns.ts
@@ -9,10 +9,7 @@ import { Context } from '../../context';
 import { FieldValue, Row } from '../../types';
 import { TransformParser } from '../parser';
 import { buildWhere } from './where';
-import {
-  mapStringToStringValidator,
-  starSingleOrMultipleColumnsValidator,
-} from './transformValidators';
+import { mapStringToStringValidator, starSingleOrMultipleColumnsValidator } from './utils';
 import { getColumnMatcher } from './helpers';
 import { TransformTable } from '../table';
